requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
Werkzeug==2.3.7
PyJWT==2.8.0

